refactor: `BaseNode` default builder & replace files
