+++ modules/App/Sources/App/Node/Nodes/CountNode.swift
@@ -1,0 +1,10 @@
+//  CountNode.swift
+//  Created by Igor Manakov on 04.03.2024.
+
+import Logic
+
+final class CountNode: BaseNode {
+    convenience init() {
+        self.init(model: Count())
+    }
+}
