+++ modules/App/Sources/App/Node/Nodes/GreaterNode.swift
@@ -1,0 +1,10 @@
+//  GreaterNode.swift
+//  Created by Vladimir Roganov on 07.03.2024
+
+import Logic
+
+final class GreaterNode: BaseNode {
+    convenience init() {
+        self.init(model: Greater())
+    }
+}
