--- modules/App/Sources/App/Node/Nodes/Join/JoinNode.swift
@@ -1,23 +1,0 @@
-//  JoinNode.swift
-//  Created by Aiur Arkhipov on 22.02.2024.
-
-import SwiftUI
-import Logic
-import Combine
-
-final class JoinNode: BaseNode {
-
-    convenience init() {
-        self.init(model: Join())
-    }
-
-    override func build(controller: LinkController, id: String) -> AnyView {
-        AnyView(BaseNodeView(
-            model: self,
-            onLinkTap: { (param) in
-                controller.link(id: id, param: param)
-            }, display: {
-                EmptyView()
-            }))
-    }
-}
